Release 2020.4

+++ GenBOE.Database/PTM/Object Scripts/Stored Procedures/deleteAttachment.proc.sql
@@ -1,6 +1,5 @@
 ﻿IF  EXISTS (SELECT 1 FROM sys.objects WHERE object_id = OBJECT_ID(N'[dbo].[deleteAttachment]') AND type in (N'P', N'PC'))
 	DROP PROCEDURE [dbo].[deleteAttachment];
-
 GO
 
 SET ANSI_NULLS ON
@@ -11,7 +10,9 @@ GO
 CREATE PROCEDURE [dbo].[deleteAttachment]
 (
 @AttachmentID int,
-@UpdateDate datetime2
+@UpdateDate datetime2,
+@ProposalId int,
+@DeleteAttachmentRecord bit
 )
 AS
 /******************************************************************************
@@ -19,7 +20,6 @@ AS
 **		Name: [deleteAttachment]
 **		Desc: Delete Attachment
 **			
-**		
 **
 **		Auth: twilson3
 **		Date: 9/7/2017
@@ -27,24 +27,22 @@ AS
 **		Change History
 *******************************************************************************
 **		Date:		Author:				Description:
-**		--------	--------			---------------------------------------
+**      --------    --------            ---------------------------------------
+**      7/1/2020	Dusan				BOEJ-4638 Post submittal attachments working with Revisioned Proposal
 *******************************************************************************/
-SET NOCOUNT ON 
-
-	IF (SELECT UpdateDate FROM [dbo].[Attachment] WHERE ID = @AttachmentID ) = @UpdateDate
+	SET NOCOUNT ON 
+	IF (SELECT UpdateDate FROM [dbo].[Attachment] WHERE ID = @AttachmentID) = @UpdateDate
 		BEGIN
-			DELETE FROM dbo.[Attachment] WHERE ID = @AttachmentID
+			DELETE FROM dbo.[ProposalsAttachments] WHERE AttachmentId = @AttachmentID AND ProposalId = @ProposalId
+			IF @DeleteAttachmentRecord = 1 
+				DELETE FROM dbo.[Attachment] WHERE ID = @AttachmentID
 		END
 	ELSE
 		BEGIN
 			DECLARE @ErrorMessage varchar (500)
-				SET @ErrorMessage =   'The Attachment with ID ' + CAST(@AttachmentID  AS varchar(10)) + ' has been updated and is out of sync with the data in your browser.  Please refresh your data.'
-			RAISERROR (
-					@ErrorMessage, -- Message text.
-			        11, -- Severity,/*Severity Changed to 11*/
-					1 -- State,
-					)
+			SET @ErrorMessage = 'The Attachment with ID ' + CAST(@AttachmentID  AS varchar(10)) + ' has been updated and is out of sync with the data in your browser.  Please refresh your data.'
+			RAISERROR (@ErrorMessage, 11, 1)
 			RETURN
-
 		END
+
 GO
